requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
EasyProcess==1.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PySocks==1.7.1
PyVirtualDisplay==3.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
xvfbwrapper==0.2.9
pandas==2.0.3
schedule==1.2.0
telebot==0.0.5
python_dotenv==1.0.0